+++ src/app/(list)/category/[category]/page.tsx
@@ -8,7 +8,7 @@ import { Suspense } from 'react'
 import { Metadata } from 'next'
 import PostFeature from '@/feature/post'
 import { toPascalCase } from '@/shared/_utils'
-import PostList from './_components/PostList'
+import PostList from '../../_components/PostList'
 import StructuredData from '@/feature/seo/_components/StructuredData'
 import AppFeature from '@/feature/application'
 import SEOFeature from '@/feature/seo'
